+++ src/components/MultiColumnMeasuresCard.tsx
@@ -10,7 +10,8 @@ import {
   Autocomplete,
   TextField,
 } from '@mui/material';
-import { matchSorter } from 'match-sorter';
+import { matchSorter, rankings } from 'match-sorter';
+import { useState } from 'react';
 import { AvailableTerm, CardDisplay, ColumnOption } from '../types/multiColumnMeasureTypes';
 
 interface TermCardProps {
@@ -36,12 +37,15 @@ function MultiColumnMeasuresCard({
   onRemoveColumn,
   onRemoveCard,
 }: TermCardProps) {
+  const [inputQueryString, setInputQueryString] = useState('');
+
   const filterOptions = (items: AvailableTerm[], { inputValue }: { inputValue: string }) =>
     matchSorter(items, inputValue, {
       keys: [
         (option) =>
           option.abbreviation ? `${option.abbreviation} - ${option.label}` : option.label,
       ],
+      threshold: rankings.CONTAINS,
     });
 
   return (
@@ -98,6 +102,14 @@ function MultiColumnMeasuresCard({
                 disableCloseOnSelect
                 data-cy={`multi-column-measures-card-${cardIndex}-columns-dropdown`}
                 options={columnOptions}
+                inputValue={inputQueryString}
+                onInputChange={(_, newInputValue, reason) => {
+                  // We want to make sure we only update the text in the dropdown
+                  // when the user types, not when the user makes a selection
+                  if (reason === 'input' || reason === 'clear') {
+                    setInputQueryString(newInputValue);
+                  }
+                }}
                 getOptionLabel={(option) => option.label}
                 getOptionDisabled={(option) => option.isPartOfCollection || false}
                 onChange={(_, newValue) => {